% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_post}
\alias{create_post}
\title{Create a Blog Post}
\usage{
create_post(title)
}
\arguments{
\item{title}{title of blog post}
}
\description{
Initializes a directory in the \code{posts} directory of the blog.
}
\keyword{internal}
